+++ examples/credential-viewer/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @learncard/credential-viewer
 
+## 0.0.23
+
+### Patch Changes
+
+-   Updated dependencies [[`69102b041690d8d161379ce9343c50a3dec0c50e`](https://github.com/learningeconomy/LearnCard/commit/69102b041690d8d161379ce9343c50a3dec0c50e)]:
+    -   @learncard/types@5.18.2
+    -   @learncard/credential-library@1.0.17
+    -   @learncard/init@2.4.12
+    -   @learncard/lca-api-plugin@2.0.1
+
 ## 0.0.22
 
 ### Patch Changes