+++ packages/plugins/did-web-plugin/CHANGELOG.md
@@ -1,5 +1,12 @@
 # did-web-plugin
 
+## 1.1.31
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.4.31
+
 ## 1.1.30
 
 ### Patch Changes